First draft of designs for tables 2 through 10
Table 1 is still incomplete, so these designs are preliminary, but it's good to share even if it has to be redone. Table 9 has not been started yet. Table 5 is the one most thoroughly worked on, as the logic for labelling the smoking columns as desired is quite in-depth. 

Includes a test case and data for when no smoking records exist in a particular month, and the most recent record has to be looked up.